refactoring: move method createInstanceOf() from AbstractDriverFactory to DefaultDriverFactory
some years ago, it was reused between several factories that used reflection: Opera, Safari, HtmlUnit etc. Nowadays none of them needs to create webdriver by reflection.
